backend/requirements.txt
# RAG Pipeline Hukum Indonesia - Requirements
# Python 3.10+ recommended

# ==== Core Framework ====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ==== Document Processing ====
pypdf>=3.17.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# ==== Text Processing ====
langchain-text-splitters>=0.0.1

# ==== Embeddings ====
sentence-transformers>=2.2.2
torch>=2.0.0

# ==== Vector Store ====
pinecone-client>=3.0.0

# ==== BM25 Search ====
rank-bm25>=0.2.2

# ==== LLM ====
llama-cpp-python>=0.2.0
huggingface-hub>=0.19.0

# ==== Utilities ====
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0

# ==== Optional: Testing ====
pytest>=7.4.0
httpx>=0.25.0

# ==== Optional: Development ====
black>=23.0.0
isort>=5.12.0
